+++ tests/functional/Extension/Attributes/LocalDataTest.php
@@ -16,6 +16,8 @@ namespace League\CommonMark\Tests\Functional\Extension\Attributes;
 
 use League\CommonMark\Environment\Environment;
 use League\CommonMark\Extension\Attributes\AttributesExtension;
+use League\CommonMark\Extension\CommonMark\CommonMarkCoreExtension;
+use League\CommonMark\Extension\GithubFlavoredMarkdownExtension;
 use League\CommonMark\MarkdownConverter;
 use League\CommonMark\Tests\Functional\AbstractLocalDataTest;
 
@@ -26,7 +28,9 @@ final class LocalDataTest extends AbstractLocalDataTest
 {
     protected function setUp(): void
     {
-        $environment = Environment::createGFMEnvironment();
+        $environment = new Environment();
+        $environment->addExtension(new CommonMarkCoreExtension());
+        $environment->addExtension(new GithubFlavoredMarkdownExtension());
         $environment->addExtension(new AttributesExtension());
         $this->converter = new MarkdownConverter($environment);
     }